Update french.xml
Good morning

A small improvement to the French.xml file in the Gup updater part.

Below are the dialog boxes depending on whether a Shift is present or not.
The title of the dialog boxes does not seem meaningful to me and I suggest giving both boxes the same title “Notepad++ Update”. It is therefore necessary to create a variable so that the translation of “Notepad++ update” is possible in cases where there is no Shift. (e.g. MSGID_NOUPDATETITLE or something like that)
For aesthetic reasons, I reversed the phrase "Go to page..., for manual shift" into "For manual shift..., go to download page. »
We would also need to create a variable to allow the “Close” button to be translated. (MSGID_CLOSEBTN for example)
For the Shift available box I removed is available in the title and I added tabs so that the double dots: are well aligned.
I also changed the text “Yes (silent)” to “Yes silently” because space is limited.
I also corrected the copyright date and version number.
Sincerely.